+++ kleidicv/src/filters/median_blur_neon.cpp
@@ -5,11 +5,14 @@
 #include "kleidicv/ctypes.h"
 #include "kleidicv/filters/filter_2d.h"
 #include "kleidicv/filters/filter_2d_5x5_neon.h"
+#include "kleidicv/filters/filter_2d_7x7_neon.h"
 #include "kleidicv/filters/median_blur.h"
 #include "kleidicv/kleidicv.h"
 #include "kleidicv/neon.h"
 #include "kleidicv/workspace/border_5x5.h"
+#include "kleidicv/workspace/border_7x7.h"
 #include "median_blur_sorting_network_5x5.h"
+#include "median_blur_sorting_network_7x7.h"
 
 namespace kleidicv::neon {
 
@@ -89,19 +92,54 @@ class MedianBlur<ScalarType, 5> : public MedianBlurBase<ScalarType> {
   }
 };  // end of class MedianBlur<ScalarType, 5>
 
+// Template for Median Blur 7x57 filters.
+template <typename ScalarType>
+class MedianBlur<ScalarType, 7> : public MedianBlurBase<ScalarType> {
+ public:
+  using SourceType = ScalarType;
+  using DestinationType = SourceType;
+  using SourceVectorType = typename VecTraits<SourceType>::VectorType;
+  using DestinationVectorType = typename VecTraits<DestinationType>::VectorType;
+  using VectorComparator =
+      typename MedianBlurBase<SourceType>::vectorized_comparator;
+  using ScalarComparator =
+      typename MedianBlurBase<SourceType>::scalar_comparator;
+
+  template <typename KernelWindowFunctor>
+  void vector_path(KernelWindowFunctor& KernelWindow,
+                   DestinationVectorType& dst) const {
+    Monostate ctx;
+    sorting_network7x7<VectorComparator>(KernelWindow, dst, ctx);
+  }
+
+  template <typename KernelWindowFunctor>
+  void scalar_path(KernelWindowFunctor& KernelWindow,
+                   DestinationType& dst) const {
+    Monostate ctx;
+    sorting_network7x7<ScalarComparator>(KernelWindow, dst, ctx);
+  }
+};  // end of class MedianBlur<ScalarType, 7>
+
 template <typename T>
 kleidicv_error_t median_blur_stripe(const T* src, size_t src_stride, T* dst,
                                     size_t dst_stride, size_t width,
                                     size_t height, size_t y_begin, size_t y_end,
-                                    size_t channels,
-                                    [[maybe_unused]] size_t kernel_width,
+                                    size_t channels, size_t kernel_width,
                                     [[maybe_unused]] size_t kernel_height,
                                     FixedBorderType border_type) {
   Rectangle rect{width, height};
   Rows<const T> src_rows{src, src_stride, channels};
   Rows<T> dst_rows{dst, dst_stride, channels};
-  MedianBlur<T, 5> median_filter;
-  Filter2D5x5<MedianBlur<T, 5>> filter{median_filter};
+  if (kernel_width == 5) {
+    MedianBlur<T, 5> median_filter;
+    Filter2D5x5<MedianBlur<T, 5>> filter{median_filter};
+    process_filter2d(rect, y_begin, y_end, src_rows, dst_rows, border_type,
+                     filter);
+    return KLEIDICV_OK;
+  }
+
+  MedianBlur<T, 7> median_filter;
+  Filter2D7x7<MedianBlur<T, 7>> filter{median_filter};
   process_filter2d(rect, y_begin, y_end, src_rows, dst_rows, border_type,
                    filter);
   return KLEIDICV_OK;